updated quickstart guide link
Erin has created a new QuickStart for getting started on the National Systems, which has more relevant info than our older 'QuickStart Guide for New Users' (which is fairly out of date now with most of our systems now offline). I'm going to be re-writing the QuickStart for New Users (and sharing with analysts for review approval), but until that's ready, we figured it would be better to point people to this new QuickStart instead. And is fine to still link to the mini-webinars as well.... not a problem that they are both referring to the national systems. I tweaked the description for the mini-webinars a bit.